refactor (unit): channge structure
